updateAt and prsima connect

+++ src/api/handlers/updates.ts
@@ -19,14 +19,22 @@ export class UpdateHandlers {
 	async createUpdate(req, res) {
 		try {
 			const { title, body, status, productId } = req.body;
+			const product = prisma.product.findUnique({
+				where: { id: productId },
+			});
+
+			if (!product) {
+				res.status(400);
+				res.json({ message: "Not your product" });
+				return;
+			}
+
 			const newUpdate = await prisma.update.create({
 				data: {
 					title,
 					body,
 					status,
-					productId,
-					created_at: new Date(),
-					updated_at: new Date(),
+					product: { connect: { id: productId } },
 				},
 			});
 
@@ -42,6 +50,29 @@ export class UpdateHandlers {
 	async updateUpdate(req, res) {
 		try {
 			const { title, body, status, productId } = req.body;
+
+			const product = await prisma.product.findMany({
+				where: {
+					belongsToId: req.user.id,
+				},
+				include: {
+					updates: true,
+				},
+			});
+
+			//TODO: remove the code below , instead filter in the query, and additionaly prevent data manipulation
+			const updates = product.reduce((allUpdates, product) => {
+				return [...allUpdates, ...product.updates];
+			}, []);
+
+			const match = updates.find((update) => update.id === req.params.id);
+
+			if (!match) {
+				res.json({ message: "Not your product" });
+
+				res.json(400);
+				return;
+			}
 
 			const updated = await prisma.update.update({
 				where: { id: req.params.id, productId },
@@ -63,8 +94,32 @@ export class UpdateHandlers {
 
 	async deleteUpdate(req, res) {
 		try {
+			const updateId = req.params.id;
+			const product = await prisma.product.findMany({
+				where: {
+					belongsToId: req.user.id,
+				},
+				include: {
+					updates: true,
+				},
+			});
+
+			//TODO: remove the code below , instead filter in the query, and additionaly prevent data manipulation
+			const updates = product.reduce((allUpdates, product) => {
+				return [...allUpdates, ...product.updates];
+			}, []);
+
+			const match = updates.find((update) => update.id === updateId);
+
+			if (!match) {
+				res.json({ message: "Not your product" });
+
+				res.json(400);
+				return;
+			}
+
 			const deleted = await prisma.update.delete({
-				where: { id: req.params.id, productId: req.body.productId },
+				where: { id: updateId },
 			});
 
 			res.json({ data: { deleted } });
